Implement#57 Revise REST API PUT; Rename class DataElementResource and other DataElement classes
