+++ eureka-common/src/main/java/edu/emory/cci/aiw/cvrg/eureka/common/comm/PhenotypeField.java
