+++ eureka-common/src/main/java/edu/emory/cci/aiw/cvrg/eureka/common/entity/PhenotypeEntity.java
